requirements.txt
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.0
huggingface-hub==0.16.4
cassio
langchain
langchain-community
langchain-google-genai
streamlit
gTTS==2.2.4 
python-dotenv
requests
numpy








